Chore: Move marketplace and plugins hooks into their page contexts
