--- packages/core/admin/admin/src/hooks/marketplace/useFetchMarketplaceProviders/index.js
@@ -1,44 +1,0 @@
-import { useNotification } from '@strapi/helper-plugin';
-import qs from 'qs';
-import { useQuery } from 'react-query';
-
-import { MARKETPLACE_API_URL } from '../constants';
-
-const fetchMarketplaceProviders = async (params = {}) => {
-  try {
-    const queryString = qs.stringify(qs.parse(params));
-    const res = await fetch(`${MARKETPLACE_API_URL}/providers?${queryString}`);
-
-    if (!res.ok) {
-      throw new Error('Failed to fetch marketplace providers.');
-    }
-
-    const data = await res.json();
-
-    return data;
-  } catch (error) {
-    console.error(error);
-  }
-
-  return null;
-};
-
-const useFetchMarketplaceProviders = (notifyLoad, params) => {
-  const toggleNotification = useNotification();
-
-  return useQuery(['list-marketplace-providers', params], () => fetchMarketplaceProviders(params), {
-    onSuccess() {
-      if (notifyLoad) {
-        notifyLoad();
-      }
-    },
-    onError() {
-      toggleNotification({
-        type: 'warning',
-        message: { id: 'notification.error', defaultMessage: 'An error occured' },
-      });
-    },
-  });
-};
-
-export default useFetchMarketplaceProviders;
