--- packages/core/admin/admin/src/hooks/useFetchEnabledPlugins/utils/api.js
@@ -1,10 +1,0 @@
-import { getFetchClient } from '@strapi/helper-plugin';
-
-const fetchEnabledPlugins = async () => {
-  const { get } = getFetchClient();
-  const { data } = await get('/admin/plugins');
-
-  return data;
-};
-
-export { fetchEnabledPlugins };
